refactor(routes): extract playback play toggle routes
