+++ docs/ARCHITECTURE_PHASE_1_ROADMAP.md
@@ -246,9 +246,10 @@ Progress:
   `/playback/state`.
 - Extracted pause/resume/toggle-pause, seek, volume, mute, and
   `/playback/state` into `app/relaytv_app/routes/playback.py`.
-- Kept `/playback/play`, `/playback/toggle`, close, play-now, resume-session,
-  temporary playback, and upload ingest playback in the aggregate router for
-  later M4 slices.
+- Extracted `/playback/play` and `/playback/toggle` into
+  `app/relaytv_app/routes/playback.py`.
+- Kept close, play-now, resume-session, temporary playback, and upload ingest
+  playback in the aggregate router for later M4 slices.
 
 ### M5: Extract Settings Router
 
@@ -369,6 +370,7 @@ Add entries here as PRs land into `codex/architecture-phase-1`.
 | 2026-06-30 | local | `codex/architecture-phase-1` | Closed M3 after confirming queue/history endpoints moved and `/now_playing/clear` remains in the playback/UI inventory group. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_queue_history_routes.py tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Begin M4 playback router planning. |
 | 2026-06-30 | local | `codex/architecture-phase-1` | Started M4 by adding HTTP-level playback route guardrails before moving playback routes. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_playback_routes.py tests/test_queue_history_routes.py tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Extract a narrow playback control slice first. |
 | 2026-06-30 | local | `codex/architecture-phase-1` | Extracted low-risk playback controls and `/playback/state` into `app/relaytv_app/routes/playback.py`. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_playback_routes.py tests/test_queue_history_routes.py tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M4 with `/playback/play` and `/playback/toggle` or close/play-now planning. |
+| 2026-06-30 | local | `codex/architecture-phase-1` | Extracted `/playback/play` and `/playback/toggle` into the playback router. | `ruff check app tests`; `PYTHONPATH=app pytest -q tests/test_playback_routes.py tests/test_queue_history_routes.py tests/test_route_inventory.py tests/test_smoke.py`; `git diff --check` | Continue M4 with close/play-now planning. |
 
 ## Open Questions
